+++ src/types/sendEmail.type.ts
@@ -3,3 +3,11 @@ export type SendEmail = {
   username: string
   group_no: string
 }
+
+export type SendDailyProblemEmail = {
+  problem_slug: string
+  difficulty: string
+  topic: string
+  group_no: string
+  email: string
+}