+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Changelog
-- 1.0.2: Feature: Automate "add group progress" cron job
-- 1.0.1: Fix: Show validation error message when the form is submitted
+
+- 1.0.3: FEAT: Automate "send daily problem email" cron job
+- 1.0.2: FEAT: Automate "add group progress" cron job
+- 1.0.1: FIX: Show validation error message when the form is submitted
 - 1.0.0: Initial release